Added dynamic graph page, controller, and banner entry.
So far, only a few of the consensus graphs have been added to
this page. It is largely a wrapper around the graph servlets.

The range controller will calculate date ranges.  This will be used to
calculate common date ranges for the current graph pages which will likely
be passed onto the image servlets. For example, it will give us the
start/end date ranges for the past X days, years, etc.

Moved date ranges controller to util package.